backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
gunicorn>=22.0.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.1
bcrypt>=4.0.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic[email]>=2.7.4
pydantic-settings>=2.3.4
redis>=5.0.6
celery>=5.4.0
pytest>=8.2.2
httpx>=0.27.0
pytest-asyncio>=0.23.7
